# AgriBridge-Connect/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgriBridge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/book:
    post:
      operationId: bookStorage
      tags: [coldStorage]
      summary: Book cold storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookStorageRequest"
      responses:
        "200":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /storage/bookings/{id}/approve:
    put:
      operationId: approveBooking
      tags: [coldStorage]
      summary: Approve or reject a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveBookingRequest"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /storage/my:
    get:
      operationId: getMyStorage
      tags: [coldStorage]
      summary: Get storage centers managed by current user
      responses:
        "200":
          description: List of storage centers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColdStorage"
    post:
      operationId: createStorage
      tags: [coldStorage]
      summary: Create a cold storage center
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorageRequest"
      responses:
        "200":
          description: Created storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColdStorage"
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/fcm-token:
    put:
      operationId: updateFcmToken
      tags: [auth]
      summary: Update FCM push notification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFcmTokenRequest"
      responses:
        "200":
          description: Token updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        location:
          type: string
        farmName:
          type: string
        bio:
          type: string
        profileImage:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        location:
          type: string
        farmName:
          type: string
        bio:
          type: string
        profileImage:
          type: string
    UpdateFcmTokenRequest:
      type: object
      properties:
        fcmToken:
          type: string
      required:
        - fcmToken
...
    ColdStorage:
      type: object
      properties:
        id:
          type: integer
        managerId:
          type: integer
        name:
          type: string
        description:
          type: string
        location:
          type: string
        totalCapacity:
          type: string
        availableCapacity:
          type: string
        pricePerKgPerDay:
          type: string
        temperature:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        available:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - managerId
        - name
        - location
        - totalCapacity
        - availableCapacity
        - pricePerKgPerDay
        - available
...
    CreateStorageRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        totalCapacity:
          type: number
        pricePerKgPerDay:
          type: number
        temperature:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
      required:
        - name
        - location
        - totalCapacity
        - pricePerKgPerDay
    Booking:
      type: object
      properties:
        id:
          type: integer
        farmerId:
          type: integer
        storageId:
          type: integer
        managerId:
          type: integer
        productName:
          type: string
        quantity:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalCost:
          type: string
        status:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - farmerId
        - storageId
        - managerId
        - productName
        - quantity
        - startDate
        - endDate
        - status
...
    BookStorageRequest:
      type: object
      properties:
        storageId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        notes:
          type: string
      required:
        - storageId
        - productName
        - quantity
        - startDate
        - endDate
    ApproveBookingRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status
...