# AgriBridge-Connect/lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriBridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  farmerId: number;
  storageId: number;
  managerId: number;
  productName: string;
  quantity: string;
  startDate: string;
  endDate: string;
  totalCost?: string;
  status: string;
  notes?: string;
  createdAt?: string;
}